+++ apps/ymir-sdl3/src/app/ui/windows/about_window.cpp
@@ -675,6 +675,7 @@ void AboutWindow::DrawAcknowledgementsTab() {
                            "Elcorsico 28, "
                            "Giovani Avelar, "
                            "Israel Jacquez, "
+                           "James Wood, "
                            "Jeff Greulich, "
                            "Joek, "
                            "Julien P, "