+++ CHANGELOG.md
@@ -8,10 +8,13 @@ Uses save state file version 12.
 
 ### New features and improvements
 
+- App: Rotate screenshots based on selected display rotation.
 - SCU: Optimize DSP operations involving 48-bit registers.
 
 ### Fixes
 
+- App: Temporarily switch from exclusive to borderless fullscreen when opening a file dialog then back to exclusive mode once the dialog is closed to avoid minimizing or resizing the window.
+- CD Block: Open tray when ejecting a disc. Fixes various cases of games that hang, freeze, crash or exhibit erratic behavior when ejecting a disc while playing. These should now properly boot back to the system's interactive shell. (#298)
 - Media: Ensure a valid Saturn disc is loaded before trying to parse the filesystem. Fixes crashes when trying to load non-Saturn discs.